refactor: code splitting
